Update preferredsoundplayer.py
A major addition in this module is a garbage collector to close all opened files, opened with c commands.  In addition to this, I rewrote how the main functions are called in that the script initializes a single instance of the windows sound player in windows, so that the aliases/sounds can be tracked.  The garbage collection is called every time a new sound is played.  It looks through a list of previous sounds, if they are no longer playing, they are closed and then removed from the garbage collection list.  So in Windows one instance of the sound player is used.  In Linux, if the point arises where the gst playbin player must be used, each sound gets its own player, to avoid internal conflicts in the gst player module.